\name{GFDrmstGUI}
\alias{GFDrmstGUI}
\title{
A Shiny app for multiple RMST-based tests
}
\description{
This function provides a shiny app for performing multiple RMST-based tests. Here, the asymptotic, groupwise bootstrap or permutation test can be applied.
}
\usage{
GFDrmstGUI()
}
\details{
Note that the calculations for the groupwise bootstrap and the permutation test may take a while, especially with stepwise extension.
}
\value{
No return value, called for side effects.
}
